Move API files directly into views
