+++ server/core/views/__init__.py
@@ -1,1 +1,23 @@
-from .api import *  # noqa
+from core.models import Class, Program
+from core.serializers import ClassSerializer
+from rest_framework.response import Response
+from rest_framework.views import APIView
+
+from .accounts import *  # noqa
+from .dashboard import *  # noqa
+from .studentreg import *  # noqa
+
+# TODO figure out object permissions for all API calls
+
+
+class ClassCatalog(APIView):
+    """
+    API endpoint that returns all classes.
+    Permissions: authenticated
+    # TODO have some publicly accessible version of this API
+    """
+
+    def get(self, request, program, edition, format=None):
+        prog = Program.objects.get(name=program, edition=edition)
+        classes = Class.objects.filter(program=prog)
+        return Response([ClassSerializer(clazz).data for clazz in classes])
